+++ internal/kubenurse/server.go
@@ -4,7 +4,7 @@ package kubenurse
 import (
 	"context"
 	"fmt"
-	"log"
+	"log/slog"
 	"net/http"
 	"os"
 	"strconv"
@@ -119,10 +119,11 @@ func New(ctx context.Context, c client.Client) (*Server, error) { //nolint:funle
 
 	if bucketsString := os.Getenv("KUBENURSE_HISTOGRAM_BUCKETS"); bucketsString != "" {
 		for _, bucketStr := range strings.Split(bucketsString, ",") {
-			bucket, e := strconv.ParseFloat(bucketStr, 64)
+			bucket, err := strconv.ParseFloat(bucketStr, 64)
 
-			if e != nil {
-				log.Fatalf("couldn't parse one of the custom histogram buckets. error:\n%v", e)
+			if err != nil {
+				slog.Error("couldn't parse one of the custom histogram buckets", "bucket", bucket, "err", err)
+				os.Exit(1)
 			}
 
 			histogramBuckets = append(histogramBuckets, bucket)
@@ -242,6 +243,8 @@ func (s *Server) Run() error {
 			}
 		}()
 	}
+
+	slog.Info("kubenurse just started")
 
 	wg.Wait()
 	close(errc)