+++ internal/servicecheck/servicecheck.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
-	"log"
+	"log/slog"
 	"net"
 	"net/http"
 	"os"
@@ -26,31 +26,11 @@ const (
 // results. Other parameters of the Checker struct need to be configured separately.
 func New(_ context.Context, cl client.Client, promRegistry *prometheus.Registry,
 	allowUnschedulable bool, cacheTTL time.Duration, durationHistogramBuckets []float64) (*Checker, error) {
-	errorCounter := prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Namespace: MetricsNamespace,
-			Name:      "errors_total",
-			Help:      "Kubenurse error counter partitioned by error type",
-		},
-		[]string{"type"},
-	)
-
-	durationHistogram := prometheus.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Namespace: MetricsNamespace,
-			Name:      "request_duration",
-			Help:      "Kubenurse request duration partitioned by target path",
-			Buckets:   durationHistogramBuckets,
-		},
-		[]string{"type"},
-	)
-
-	promRegistry.MustRegister(errorCounter, durationHistogram)
-
 	// setup http transport
 	tlsConfig, err := generateTLSConfig(os.Getenv("KUBENURSE_EXTRA_CA"))
 	if err != nil {
-		log.Printf("cannot generate tlsConfig with KUBENURSE_EXTRA_CA: %s", err)
+		slog.Error("cannot generate tlsConfig with provided KUBENURSE_EXTRA_CA. Continuing with default tlsConfig",
+			"KUBENURSE_EXTRA_CA", os.Getenv("KUBENURSE_EXTRA_CA"), "err", err)
 
 		tlsConfig = &tls.Config{MinVersion: tls.VersionTLS12}
 	}
@@ -82,40 +62,26 @@ func New(_ context.Context, cl client.Client, promRegistry *prometheus.Registry,
 		client:             cl,
 		httpClient:         httpClient,
 		cacheTTL:           cacheTTL,
-		errorCounter:       errorCounter,
-		durationHistogram:  durationHistogram,
 		stop:               make(chan struct{}),
 	}, nil
 }
 
 // Run runs all servicechecks and returns the result togeter with a boolean which indicates success. The cache
 // is respected.
-func (c *Checker) Run() (Result, bool) {
-	var (
-		haserr bool
-		err    error
-	)
-
+func (c *Checker) Run() {
 	// Run Checks
 	res := Result{}
 
-	res.APIServerDirect, err = c.measure(c.APIServerDirect, "api_server_direct")
-	haserr = haserr || (err != nil)
-
-	res.APIServerDNS, err = c.measure(c.APIServerDNS, "api_server_dns")
-	haserr = haserr || (err != nil)
-
-	res.MeIngress, err = c.measure(c.MeIngress, "me_ingress")
-	haserr = haserr || (err != nil)
-
-	res.MeService, err = c.measure(c.MeService, "me_service")
-	haserr = haserr || (err != nil)
+	res.APIServerDirect = c.measure(c.APIServerDirect, "api_server_direct")
+	res.APIServerDNS = c.measure(c.APIServerDNS, "api_server_dns")
+	res.MeIngress = c.measure(c.MeIngress, "me_ingress")
+	res.MeService = c.measure(c.MeService, "me_service")
 
 	if c.SkipCheckNeighbourhood {
 		res.NeighbourhoodState = skippedStr
 	} else {
+		var err error
 		res.Neighbourhood, err = c.GetNeighbours(context.Background(), c.KubenurseNamespace, c.NeighbourFilter)
-		haserr = haserr || (err != nil)
 
 		// Neighbourhood special error treating
 		if err != nil {
@@ -130,8 +96,6 @@ func (c *Checker) Run() (Result, bool) {
 
 	// Cache result (used for /alive handler)
 	c.LastCheckResult = &res
-
-	return res, haserr
 }
 
 // RunScheduled runs the checks in the specified interval which can be used to keep the metrics up-to-date. This
@@ -196,23 +160,13 @@ func (c *Checker) MeService(ctx context.Context) (string, error) {
 }
 
 // measure implements metric collections for the check
-func (c *Checker) measure(check Check, label string) (string, error) {
-	start := time.Now()
-
+func (c *Checker) measure(check Check, requestType string) string {
 	// Add our label (check type) to the context so our http tracer can annotate
 	// metrics and errors based with the label
-	ctx := context.WithValue(context.Background(), kubenurseTypeKey{}, label)
+	ctx := context.WithValue(context.Background(), kubenurseTypeKey{}, requestType)
 
 	// Execute check
-	res, err := check(ctx)
+	res, _ := check(ctx) // this error is ignored as it is already logged in httptrace
 
-	// Process metrics
-	c.durationHistogram.WithLabelValues(label).Observe(time.Since(start).Seconds())
-
-	if err != nil {
-		log.Printf("failed request for %s with %v", label, err)
-		c.errorCounter.WithLabelValues(label).Inc()
-	}
-
-	return res, err
+	return res
 }